+++ src/pageEditor/tabs/editTab/editorNodeLayout/usePipelineNodes/useMapPipelineToNodes.ts
@@ -15,24 +15,19 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import useTypedBrickMap from "@/bricks/hooks/useTypedBrickMap";
 import { type BrickPipeline, type PipelineFlavor } from "@/bricks/types";
 import { PIPELINE_BRICKS_FIELD_NAME } from "@/pageEditor/consts";
 import {
   type MapOutput,
   type EditorNodeProps,
 } from "@/pageEditor/tabs/editTab/editorNodeLayout/usePipelineNodes/types";
+import { useGetTraceHandling } from "@/pageEditor/tabs/editTab/editorNodeLayout/usePipelineNodes/useGetTraceHandling";
+import { useGetLastBrickHandling } from "@/pageEditor/tabs/editTab/editorNodeLayout/usePipelineNodes/useLastBrickHandling";
 import { type MapBrickToNodesArgs } from "@/pageEditor/tabs/editTab/editorNodeLayout/usePipelineNodes/useMapBrickToNodes";
-import { BrickTypes } from "@/runtime/runtimeTypes";
-import { type TraceRecord } from "@/telemetry/trace";
-import {
-  filterTracesByCall,
-  getLatestBrickCall,
-} from "@/telemetry/traceHelpers";
 import { type Branch } from "@/types/runtimeTypes";
 import { useCallback } from "react";
 
-type MapPipelineToNodesArgs = {
+export type MapPipelineToNodesArgs = {
   pipeline: BrickPipeline;
   flavor: PipelineFlavor;
   pipelinePath?: string;
@@ -46,13 +41,13 @@ type MapPipelineToNodesArgs = {
    * @since 1.8.4
    */
   latestParentCall?: Branch[];
-  traces: TraceRecord[];
 };
 
 export function useMapPipelineToNodes(
   mapBrickToNodes: (args: MapBrickToNodesArgs) => MapOutput,
 ): (args: MapPipelineToNodesArgs) => MapOutput {
-  const { data: allBricks } = useTypedBrickMap();
+  const getTraceHandling = useGetTraceHandling();
+  const getLastBrickHandling = useGetLastBrickHandling();
 
   return useCallback(
     ({
@@ -63,28 +58,16 @@ export function useMapPipelineToNodes(
       isParentActive = false,
       isAncestorActive = false,
       latestParentCall,
-      traces,
     }: MapPipelineToNodesArgs) => {
       const isRootPipeline = pipelinePath === PIPELINE_BRICKS_FIELD_NAME;
-      const lastIndex = pipeline.length - 1;
-      const lastBrickId = pipeline.at(lastIndex)?.id;
-      const lastBrick = lastBrickId ? allBricks?.get(lastBrickId) : undefined;
-      const showAppend =
-        !lastBrick?.block || lastBrick.type !== BrickTypes.RENDERER;
+      const { lastIndex, showAppend } = getLastBrickHandling(pipeline);
+
+      const latestPipelineCall = getTraceHandling({
+        pipeline,
+        latestParentCall,
+      });
+
       const nodes: EditorNodeProps[] = [];
-
-      // Determine which execution of the pipeline to show. Currently, getting the latest execution
-      let latestPipelineCall: Branch[] | undefined;
-      if (pipeline.length > 0) {
-        // Pass [] as default to include all traces
-        const latestTraces = filterTracesByCall(traces, latestParentCall ?? []);
-        // Use first brick in pipeline to determine the latest run
-        latestPipelineCall = getLatestBrickCall(
-          latestTraces,
-          pipeline[0]?.instanceId,
-        )?.branches;
-      }
-
       let modComponentHasTraces = false;
 
       for (const [index, brickConfig] of pipeline.entries()) {
@@ -115,6 +98,6 @@ export function useMapPipelineToNodes(
         modComponentHasTraces,
       };
     },
-    [allBricks, mapBrickToNodes],
+    [getLastBrickHandling, getTraceHandling, mapBrickToNodes],
   );
 }
